+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Changes that happened in releases
 * Fix dynamic fetch support for `MULTISET` fetched mappings
 * Fix annotation processor code generation issue with type use annotations on primitive types
 * Fix NPE on cascade delete of unmapped `null` one-to-one association
+* Throw `OptimisticLockException` if view can't be found for entity view lifecycle listener
 
 ### Backwards-incompatible changes